improved windows prepare while usingg it  on new system, fixed a missing dll

+++ bash/linux/download_libraries.sh
@@ -6,19 +6,16 @@ sdl_url="https://github.com/libsdl-org/SDL/releases/download/release-2.30.2/SDL2
 sdl_zip="build/sdl.zip"
 sdl_path="build/linux-sdl"
 sdl_source="$sdl_path/SDL2-2.30.2" # SDL2-2.0.14"
-sdl_dest=$sdl_path
 
 sdl_image_url="https://github.com/libsdl-org/SDL_image/releases/download/release-2.8.2/SDL2_image-2.8.2.zip"
 sdl_image_zip="build/sdl_image.zip"
 sdl_image_path="build/linux-sdl_image"
 sdl_image_source="$sdl_image_path/SDL2_image-2.8.2" # 2.0.5"
-sdl_image_dest=$sdl_image_path
 
 sdl_mixer_url="https://github.com/libsdl-org/SDL_mixer/releases/download/release-2.8.0/SDL2_mixer-2.8.0.zip"
 sdl_mixer_zip="build/sdl_mixer.zip"
 sdl_mixer_path="build/linux-sdl_mixer"
 sdl_mixer_source="$sdl_mixer_path/SDL2_mixer-2.8.0" # 2.0.4"
-sdl_mixer_dest=$sdl_mixer_path
 
 glew_url="https://github.com/nigels-com/glew/releases/download/glew-2.2.0/glew-2.2.0.zip"
 # Set the directory to extract SDL2 to
@@ -43,19 +40,19 @@ fi
 
 if [ ! -f lib/libGLEW.a ]; then
     # Download glew
-    wget -O "$glew_zip" "$glew_url"
-    unzip -q "$glew_zip" -d "$glew_path"
-    rm "$glew_zip"
-    # remove folder build/glew/glew-2.2.0
-    mv "$glew_source"/* "$glew_dest"
-    rmdir "$glew_source"
-
-    # remove folder build/glew/include/GL
-    glew_source2="build/glew/include/GL"
-    glew_dest2="build/glew/include"
-    mv "$glew_source2"/* "$glew_dest2"
-    rmdir "$glew_source2"
-
+    if [ ! -d $glew_path ]; then
+        wget -O "$glew_zip" "$glew_url"
+        unzip -q "$glew_zip" -d "$glew_path"
+        rm "$glew_zip"
+        # remove folder build/glew/glew-2.2.0
+        mv "$glew_source"/* "$glew_dest"
+        rmdir "$glew_source"
+        # remove folder build/glew/include/GL
+        glew_source2="build/glew/include/GL"
+        glew_dest2="build/glew/include"
+        mv "$glew_source2"/* "$glew_dest2"
+        rmdir "$glew_source2"
+    fi
     if [ ! -f $glew_path/lib/libGLEW.a ]; then
         cd $glew_path
         make
@@ -66,12 +63,14 @@ fi
 
 if [ ! -f lib/libSDL2.la ]; then
     # download sdl2
-    wget -O "$sdl_zip" "$sdl_url"
-    unzip -q "$sdl_zip" -d
-    rm "$sdl_zip"
-    # remove folder build/sdl/SDL2-2.0.14"$sdl_path"
-    mv "$sdl_source"/* "$sdl_dest"
-    rmdir "$sdl_source"
+    if [ ! -d $sdl_path ]; then
+        wget -O "$sdl_zip" "$sdl_url"
+        unzip -q "$sdl_zip" -d "$sdl_path"
+        rm "$sdl_zip"
+        # remove folder build/sdl/SDL2-2.0.14"$sdl_path"
+        mv "$sdl_source"/* "$sdl_path"
+        rmdir "$sdl_source"
+    fi
     if [ ! -f $sdl_path/build/libSDL2.la ]; then
         cd $sdl_path
         ./configure
@@ -87,7 +86,7 @@ if [ ! -f "lib/libSDL2_image.la" ]; then
     unzip -q "$sdl_image_zip" -d "$sdl_image_path"
     rm "$sdl_image_zip"
     # remove folder build/sdl_image/SDL2-2.0.14
-    mv "$sdl_image_source"/* "$sdl_image_dest"
+    mv "$sdl_image_source"/* "$sdl_image_path"
     rmdir "$sdl_image_source"
     if [ ! -f $sdl_image_path/libSDL2_image.la ]; then
         cd $sdl_image_path
@@ -104,7 +103,7 @@ if [ ! -f "lib/libSDL2_mixer.la" ]; then
     unzip -q "$sdl_mixer_zip" -d "$sdl_mixer_path"
     rm "$sdl_mixer_zip"
     # remove folder build/sdl/SDL2-2.0.14
-    mv "$sdl_mixer_source"/* "$sdl_mixer_dest"
+    mv "$sdl_mixer_source"/* "$sdl_mixer_path"
     rmdir "$sdl_mixer_source"
     if [ ! -f $sdl_mixer_path/build/libSDL2_mixer.la ]; then
         cd $sdl_mixer_path
